Add .csv support and rewrite in OOP
Code is rewritten with OOP mentality. Arbitrary number of defenders can
be added now but the attacker number is still set to one. In future, an
upgrade will be done to simulation so we can add multiple attackers and
update the logic so defenders can choose which attacker they want to chase,
for shortest interception time for all targets.

Attacker flights can be sourced from .csv files. Defenders save their
flight log when they catch the attacker. Both cases use the same
.csv format so you can check the example .csv files.